requirements.txt
hikari == 2.0.0.dev122
hikari-crescent == 0.6.6
crescent-ext-docstrings == 0.2.2
hikari-miru == 4.1.1